requirements.txt
PySide6>=6.6.0
httpx[http2]>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pyinstaller>=6.0.0
